dtype

+++ src/diffusers/models/embeddings.py
@@ -749,7 +749,7 @@ class CombinedTimestepSizeEmbeddings(nn.Module):
 
         current_batch_size, dims = size.shape[0], size.shape[1]
         size = size.reshape(-1)
-        size_freq = self.additional_condition_proj(size)
+        size_freq = self.additional_condition_proj(size).to(size.dtype)
         # size_freq = get_timestep_embedding(
         #     size, 256, downscale_freq_shift=0, flip_sin_to_cos=True
         # ).to(size.dtype)
